Added an abstract class GenericContact for Contact entity
